Update README.md
This project was a bit difficult because the tutorial was based on linux and we had to use GitHub.
The challenging part was that while the tutorial used linuxconfig to save their files, it was more difficult for me to save because the username on the git bash was confusing and wouldn't let me save properly. While I followed the tutorial, my backup.sh doesn't show the results as the tutorial.